Update the xattr_set() interface for the 5.12 user_namespace API changes
Some more additions that update the ZPL_XATTR_SET_WRAPPER() macro such
that it will build interface functions that adopt the struct
user_namespace* argument as the second arg, if that is the
implementation required in the kernel (determined by configure). This
was added to support idmapped FS mounts in Linux 5.12.

Signed-off-by: Coleman Kane <ckane@colemankane.org>